improved

+++ events/on_messages.py
@@ -4,34 +4,21 @@ from bot_utilities.owner_utils import *
 from bot_utilities.ai_utils import generate_response_msg
 
 def on_messages(bot, cmd_list , member_histories_msg, mongodb):
+
     @bot.event
     async def on_message(message):
-        if message.guild is None:
-            return
-
-        # delete shapes commands
-        if message.content == "hello im LuminaryAI. @ me to talk w me or DM me." and message.author == bot.user:
-            await message.delete()
-            return
-        elif message.content.startswith("there are 4 ways you can interact with me:") and message.author == bot.user:
-            await message.delete()
-        elif message.content.startswith("hello im LuminaryAI. to start chatting, just tag me") and message.author == bot.user:
-            await message.delete()
-        elif message.content.startswith("uhh my head hurts") and message.author == bot.user:
-            await message.delete()
-        
+        if message.guild is None: return
+        if message.author == bot.user: return
 
         if message.author == bot.user:
-            return
-
+            if message.content.startswith("hello im LuminaryAI. @ me to talk w me or DM me."):await message.delete(); return
+            elif message.content.startswith("there are 4 ways you can interact with me:"):await message.delete(); return
+            elif message.content.startswith("hello im LuminaryAI. to start chatting, just tag me"):await message.delete(); return
+            elif message.content.startswith("uhh my head hurts"): await message.delete(); return
 
         if message.content.startswith(tuple(cmd_list)):
             if not await check_blist_msg(message, mongodb):
                 await bot.process_commands(message)
-
-
-
-
 
         elif await getdb("ai-channels", message.channel.id, mongodb) == "found":
             if await check_blist_msg(message, mongodb): return
@@ -60,9 +47,6 @@ def on_messages(bot, cmd_list , member_histories_msg, mongodb):
             )
             answer_generated.set_footer(text="Thanks for using LuminaryAI!", icon_url=bot.user.avatar.url)
             await answer.edit(embed=answer_generated)
-
-
-
 
         elif not any(message.content.startswith(prefix) for prefix in bot.command_prefix):
             return
